Refactor engine constants to resolve circular imports
